+++ src/pcobra/cobra/hub/__init__.py
@@ -1,0 +1,13 @@
+"""Infraestructura de repositorios CobraHub."""
+
+from pcobra.cobra.hub.repository import (
+    DownloadedPackage,
+    HttpCobraHubRepository,
+    PackageRepository,
+)
+
+__all__ = [
+    "DownloadedPackage",
+    "HttpCobraHubRepository",
+    "PackageRepository",
+]
